+++ l10n_es_pos_oca/__manifest__.py
@@ -9,16 +9,12 @@
     "Odoo Community Association (OCA)",
     "website": "https://github.com/OCA/l10n-spain",
     "license": "AGPL-3",
-    "version": "16.0.1.0.2",
-    "depends": ["point_of_sale"],
+    "version": "17.0.1.0.0",
+    "depends": ["point_of_sale", "l10n_es"],
     "data": ["views/pos_views.xml", "views/res_config_settings_views.xml"],
     "assets": {
-        "point_of_sale.assets": [
-            "l10n_es_pos/static/src/xml/pos.xml",
-            "l10n_es_pos/static/src/xml/Screens/OrderManagementScreen/TicketScreen.xml",
-            "l10n_es_pos/static/src/js/PaymentScreen.js",
-            "l10n_es_pos/static/src/js/models.js",
-            "l10n_es_pos/static/src/js/Screens/TicketScreen/TicketScreen.js",
+        "point_of_sale._assets_pos": [
+            "l10n_es_pos_oca/static/src/**/*",
         ],
     },
     "installable": True,
