+++ l10n_es_pos_oca/static/src/app/screens/receipt_screen/receipt/receipt_header/receipt_header.xml
@@ -1,0 +1,42 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<templates id="template" xml:space="preserve">
+    <t
+        t-name="OrderReceipt"
+        t-inherit="point_of_sale.ReceiptHeader"
+        t-inherit-mode="extension"
+    >
+        <xpath
+            expr="//div[@t-if='props.data.company.partner_id?.[1]']"
+            position="before"
+        >
+            <t t-if="props.data.is_simplified_config and !props.data.to_invoice">
+                <div>Simplified invoice</div>
+                <div
+                    class="simplified-invoice-number"
+                    t-esc="props.data.l10n_es_unique_id"
+                />
+            </t>
+        </xpath>
+        <xpath
+            expr="//div[@t-if='props.data.company.partner_id?.[1]']"
+            position="after"
+        >
+            <div t-if="props.data.company.street" t-esc="props.data.company.street" />
+            <div t-if="props.data.company.zip" t-esc="props.data.company.zip" />
+            <div t-if="props.data.company.city" t-esc="props.data.company.city" />
+            <div t-if="props.data.company.state_id">(<t
+                    t-esc="props.data.company.state_id[1]"
+                />)</div>
+        </xpath>
+        <xpath expr="//div[hasclass('pos-receipt-contact')]" position="inside">
+            <t t-set="partner" t-value="props.data.partner" />
+            <t t-if="partner and partner.id !== props.data.simplified_partner_id">
+                <div>Customer: <t t-esc="partner.name" /></div>
+                <div t-if="partner.vat"><t
+                        t-esc="props.data.company.country?.vat_label || 'Tax ID'"
+                    />: <t t-esc="partner.vat" /></div>
+                <div t-if="partner.address" t-esc="partner.address" />
+            </t>
+        </xpath>
+    </t>
+</templates>
